requirements.txt
-e .
aws_cdk.core
aws_cdk.aws_ssm
aws_cdk.aws_secretsmanager
aws_cdk.aws_iam
